requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
ipykernel
ipywidgets
torch
pandas
scikit-learn
matplotlib
seaborn
datasets
transformers>=4.50.0
Pillow
trl
wandb
evaluate
accelerate
python-dotenv
peft
openai
tenacity
google-generativeai
tiktoken
anthropic
together
reka-api>=2.0.0
zhipuai
ai21>=2.13.0
azure.identity
azure-ai-inference
cohere
mistralai
einops
deepspeed
spacy
sentence-transformers
benepar
spacy_syllables
black